fix(checker): split default type boundary helpers
